budgetcli/requirements.txt
pandas>=2.0.0
fpdf2>=2.7.0
pytest>=7.0.0
